Prevent showing refund icon without refund guarantee
